+++ examples/data-sources/auth0_event_stream/data-source.tf
@@ -1,0 +1,4 @@
+# An Auth0 Event Stream loaded using its ID.
+data "auth0_event_stream" "test" {
+  id = "est_XXXXXXXXXXXXXXX"
+}