+++ internal/auth0/client/resource.go
@@ -19,7 +19,7 @@ import (
 
 // ValidAppTypes contains all valid values for client app_type.
 var ValidAppTypes = []string{
-	"native", "spa", "regular_web", "non_interactive", "rms",
+	"native", "spa", "regular_web", "non_interactive", "resource_server", "rms",
 	"box", "cloudbees", "concur", "dropbox", "mscrm", "echosign",
 	"egnyte", "newrelic", "office365", "salesforce", "sentry",
 	"sharepoint", "slack", "springcm", "sso_integration", "zendesk", "zoom",
@@ -67,7 +67,7 @@ func NewResource() *schema.Resource {
 				Optional:     true,
 				ValidateFunc: validation.StringInSlice(ValidAppTypes, false),
 				Description: "Type of application the client represents. Possible values are: `native`, `spa`, " +
-					"`regular_web`, `non_interactive`, `sso_integration`. Specific SSO integrations types accepted " +
+					"`regular_web`, `non_interactive`, `resource_server`,`sso_integration`. Specific SSO integrations types accepted " +
 					"as well are: `rms`, `box`, `cloudbees`, `concur`, `dropbox`, `mscrm`, `echosign`, `egnyte`, " +
 					"`newrelic`, `office365`, `salesforce`, `sentry`, `sharepoint`, `slack`, `springcm`, `zendesk`, " +
 					"`zoom`.",
@@ -132,6 +132,21 @@ func NewResource() *schema.Resource {
 				Optional:    true,
 				Description: "Types of grants that this client is authorized to use.",
 			},
+			"async_approval_notification_channels": {
+				Type: schema.TypeList,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+					ValidateFunc: validation.StringInSlice([]string{
+						"guardian-push",
+						"email",
+					}, false),
+				},
+				Optional: true,
+				Computed: true,
+				Description: "List of notification channels enabled for CIBA (Client-Initiated Backchannel Authentication) requests initiated by this client. " +
+					"Valid values are `guardian-push` and `email`. The order is significant as this is the order in which notification channels will be evaluated. " +
+					"Defaults to `[\"guardian-push\"]` if not specified.",
+			},
 			"organization_usage": {
 				Type:     schema.TypeString,
 				Optional: true,
@@ -149,6 +164,21 @@ func NewResource() *schema.Resource {
 				}, false),
 				Description: "Defines how to proceed during an authentication transaction when " +
 					"`organization_usage = \"require\"`. Can be `no_prompt` (default), `pre_login_prompt` or  `post_login_prompt`.",
+			},
+			"organization_discovery_methods": {
+				Type: schema.TypeList,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+					ValidateFunc: validation.StringInSlice([]string{
+						"email", "organization_name",
+					}, false),
+				},
+				Optional: true,
+				Description: "Methods for discovering organizations during the pre_login_prompt. " +
+					"Can include `email` (allows users to find their organization by entering their email address) " +
+					"and/or `organization_name` (requires users to enter the organization name directly). " +
+					"These methods can be combined. Setting this property requires that " +
+					"`organization_require_behavior` is set to `pre_login_prompt`.",
 			},
 			"allowed_origins": {
 				Type:     schema.TypeList,
@@ -1476,10 +1506,38 @@ func NewResource() *schema.Resource {
 							Computed:    true,
 							Description: "Indicates whether the application is allowed to use a refresh token when using a session_transfer_token session.",
 						},
+						"enforce_cascade_revocation": {
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Computed:    true,
+							Description: "Indicates whether Refresh Tokens created during a native-to-web session are tied to that session's lifetime. This determines if such refresh tokens should be automatically revoked when their corresponding sessions are.",
+						},
+						"enforce_online_refresh_tokens": {
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Computed:    true,
+							Description: "Indicates whether revoking the parent Refresh Token that initiated a Native to Web flow and was used to issue a Session Transfer Token should trigger a cascade revocation affecting its dependent child entities.",
+						},
 					},
 				},
 			},
 			"token_quota": commons.TokenQuotaSchema(),
+			"skip_non_verifiable_callback_uri_confirmation_prompt": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				Description:  "Indicates whether the confirmation prompt appears when using non-verifiable callback URIs. Set to true to skip the prompt, false to show it, or null to unset. Accepts (true/false/null) or (\"true\"/\"false\"/\"null\") ",
+				ValidateFunc: validation.StringInSlice([]string{"true", "false", "null"}, false),
+				DiffSuppressFunc: func(_, o, n string, _ *schema.ResourceData) bool {
+					return (o == "null" && n == "") || o == n
+				},
+			},
+			"resource_server_identifier": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Description: "The identifier of a resource server that client is associated with" +
+					"This property can be sent only when app_type=resource_server." +
+					"This property can not be changed, once the client is created.",
+			},
 		},
 	}
 }