+++ internal/auth0/tenant/resource.go
@@ -450,6 +450,15 @@ func NewResource() *schema.Resource {
 					},
 				},
 			},
+			"skip_non_verifiable_callback_uri_confirmation_prompt": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				Description:  "Indicates whether the confirmation prompt appears when using non-verifiable callback URIs. Set to true to skip the prompt, false to show it, or null to unset. Accepts (true/false/null) or (\"true\"/\"false\"/\"null\") ",
+				ValidateFunc: validation.StringInSlice([]string{"true", "false", "null"}, false),
+				DiffSuppressFunc: func(_, o, n string, _ *schema.ResourceData) bool {
+					return (o == "null" && n == "") || o == n
+				},
+			},
 		},
 	}
 }